feat: implement emails list interface with animated table components and shared data utilities
